Renamed ceph package to ocs. Moved contents of oc package to ocs and removed oc.
